feat(helper): create helper function to format fields
