+++ apps/api/src/controllers/applicationController.ts
@@ -23,9 +23,9 @@ import { getDbInstance } from '@/db/index.js';
 import logger from '@/logger.js';
 import { type ApplicationListRequest } from '@/routes/types.js';
 import { applicationSvc } from '@/service/applicationService.js';
-import { type ApplicationContentUpdates, type ApplicationRecord, type ApplicationService } from '@/service/types.js';
+import { type ApplicationRecord, type ApplicationService } from '@/service/types.js';
 import { failure, success, type AsyncResult } from '@/utils/results.js';
-import { aliasApplicationRecord } from '@/utils/routes.js';
+import { aliasApplicationContentsRecord, aliasApplicationRecord } from '@/utils/routes.js';
 import { type UpdateEditApplicationRequest } from '@pcgl-daco/validation';
 import { ApplicationStateEvents, ApplicationStateManager } from './stateManager.js';
 
@@ -71,38 +71,9 @@ export const editApplication = async ({ id, update }: { id: number; update: Upda
 		return failure(message);
 	}
 
-	const formatedUpdate: ApplicationContentUpdates = {
-		applicant_first_name: update.applicantFirstname,
-		applicant_middle_name: update.applicantMiddlename,
-		applicant_last_name: update.applicantLastname,
-		applicant_institutional_email: update.applicantInstitutionalEmail,
-		applicant_position_title: update.applicantPositionTitle,
-		applicant_primary_affiliation: update.applicantPrimaryAffiliation,
-		applicant_profile_url: update.applicantProfileUrl,
-		applicant_suffix: update.applicantSuffix,
-		applicant_title: update.applicantTitle,
-		institutional_rep_first_name: update.institutionalRepFirstname,
-		institutional_rep_middle_name: update.institutionalRepMiddlename,
-		institutional_rep_last_name: update.institutionalRepLastname,
-		institutional_rep_title: update.institutionalRepTitle,
-		institutional_rep_position_title: update.institutionalRepPositionTitle,
-		institutional_rep_suffix: update.institutionalRepSuffix,
-		institutional_rep_email: update.institutionalRepEmail,
-		institutional_rep_primary_affiliation: update.institutionalRepPrimaryAffiliation,
-		institutional_rep_profile_url: update.institutionalRepProfileUrl,
-		institution_building: update.institutionBuilding,
-		institution_city: update.institutionCity,
-		institution_country: update.institutionCountry,
-		institution_postal_code: update.institutionPostalCode,
-		institution_state: update.institutionState,
-		institution_street_address: update.institutionStreetAddress,
-		project_aims: update.projectAims,
-		project_methodology: update.projectMethodology,
-		project_summary: update.projectSummary,
-		project_title: update.projectTitle,
-		project_website: update.projectWebsite,
-	};
-	return await applicationRepo.editApplication({ id, update: formatedUpdate });
+	const data = aliasApplicationContentsRecord(update);
+
+	return await applicationRepo.editApplication({ id, update: data });
 };
 
 /**
